pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "openmemory"
version = "0.1.0"
requires-python = ">=3.12"
description = "SIGMA: Self-evolving Intelligence for Graphs, Memory & Analysis"

dependencies = [
    # Core Framework
    "boto3>=1.34.0",
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "sqlalchemy>=1.4.0",
    "python-dotenv>=0.19.0",
    "alembic>=1.7.0",
    "psycopg2-binary>=2.9.0",
    "python-multipart>=0.0.5",
    "fastapi-pagination>=0.12.0",
    
    # MCP Server
    "mcp[cli]>=1.3.0",
    
    # Memory & Embeddings
    "mem0ai>=0.1.92",
    "openai>=1.40.0",
    
    # LLM Providers
    "anthropic>=0.51.0",
    "ollama>=0.4.8",
    
    # HTTP & Utils
    "httpx>=0.24.0",
    "tenacity>=9.1.2",
    "apscheduler>=3.10.0",
    
    # SIGMA Phase 1: Knowledge Graph (Graphiti + Neo4j)
    "graphiti-core>=0.5.0",
    "neo4j>=5.0.0",
    
    # SIGMA Phase 2: Git Integration
    "gitpython>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
]

# Git Integration (Phase 2 - uncomment when ready)
git = [
    "gitpython>=3.1.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
]
